Optimize view Versions by completely rewriting it
The view was running slow since years, because it joined more and
more of the same table (packages for new Slackware versions) thus
generating billions of rows. The last version including packages
from Slackware 10.2 to Slackware 14.2 ran for about 12.5 seconds
on the server. So to optimize its performance rewrite the view as
union of queries selecting only latest packages for each Slackware
version. The view now runs for about 0.016 seconds.