Bug 117: do not require absolute path to package file
With the modification to change working directory in BEGIN block, so
libraries could be loaded, commands against filepath are failing with
'No such file or directory' error. Since the script is normally run
outside SlackPack's working tree (i.e. in a repository) it is better
to add script's path to INC in order to load used modules like it is
already done in other scripts (e.g. Build.PL), instead of changing
script's working directory in BEGIN block.

See https://perlmaven.com/how-to-change-inc-to-find-perl-modules-in-non-standard-locations ,
    https://perlmaven.com/begin